Share bookmark component between lists
